core: split dialect interfaces into own files
