+++ xdsl/dialects/arith.py
@@ -4,7 +4,9 @@ import abc
 from collections.abc import Mapping, Sequence
 from typing import ClassVar, Literal, cast
 
-from xdsl.dialect_interfaces import ConstantMaterializationInterface
+from xdsl.dialect_interfaces.constant_materialization import (
+    ConstantMaterializationInterface,
+)
 from xdsl.dialects.builtin import (
     AnyFloat,
     AnyFloatConstr,
